+++ src/serial.rs
@@ -107,6 +107,7 @@ usart_pins! {
     feature = "stm32f030xc",
     feature = "stm32f042",
     feature = "stm32f070",
+    feature = "stm32f072",
 ))]
 usart_pins! {
     USART2 => {
@@ -114,14 +115,18 @@ usart_pins! {
         rx => [gpioa::PA3<Alternate<AF1>>, gpioa::PA15<Alternate<AF1>>],
     }
 }
-#[cfg(feature = "stm32f072Vx")]
+#[cfg(feature = "stm32f072")]
 usart_pins! {
     USART2 => {
         tx => [gpiod::PD5<Alternate<AF0>>],
         rx => [gpiod::PD6<Alternate<AF0>>],
     }
 }
-#[cfg(any(feature = "stm32f030xc", feature = "stm32f070xb"))]
+#[cfg(any(
+    feature = "stm32f030xc",
+    feature = "stm32f070xb",
+    feature = "stm32f072"
+))]
 usart_pins! {
     USART3 => {
         // According to the datasheet PB10 is both tx and rx, but in stm32cubemx it's only tx
@@ -136,39 +141,10 @@ usart_pins! {
 #[cfg(feature = "stm32f072")]
 usart_pins! {
     USART3 => {
-        tx => [gpiob::PB10<Alternate<AF4>>],
-        rx => [gpiob::PB11<Alternate<AF4>>],
-    }
-}
-#[cfg(any(feature = "stm32f072Vx", feature = "stm32f072Rx"))]
-usart_pins! {
-    USART3 => {
-        tx => [gpioc::PC4<Alternate<AF1>>, gpioc::PC10<Alternate<AF1>>],
-        rx => [gpioc::PC5<Alternate<AF1>>, gpioc::PC11<Alternate<AF1>>],
-    }
-}
-#[cfg(feature = "stm32f072Vx")]
-usart_pins! {
-    USART3 => {
         tx => [gpiod::PD8<Alternate<AF0>>],
         rx => [gpiod::PD9<Alternate<AF0>>],
     }
 }
-#[cfg(feature = "stm32f072")]
-usart_pins! {
-    USART4 => {
-        tx => [gpioa::PA0<Alternate<AF4>>],
-        rx => [gpioa::PA1<Alternate<AF4>>],
-    }
-}
-#[cfg(any(feature = "stm32f072Vx", feature = "stm32f072Rx"))]
-usart_pins! {
-    USART4 => {
-        tx => [gpioc::PC10<Alternate<AF0>>],
-        rx => [gpioc::PC11<Alternate<AF0>>],
-    }
-}
-
 #[cfg(feature = "stm32f030xc")]
 usart_pins! {
     USART5 => {